Code cleanup, clippy fixes, and kernel CI

+++ crates/bedrock-vmx/src/decoder.rs
@@ -158,7 +158,7 @@ pub fn decode_instruction(bytes: &[u8]) -> Result<DecodedInstruction, DecodeErro
             let rm = (modrm & 0x07) | (if rex.b { 0x08 } else { 0 });
 
             // Calculate ModR/M length (displacement bytes)
-            let modrm_len = modrm_displacement_length(mod_bits, rm as u8 & 0x07, &bytes[pos..])?;
+            let modrm_len = modrm_displacement_length(mod_bits, rm & 0x07, &bytes[pos..])?;
             pos += modrm_len;
 
             // mod=11 means register-to-register, not memory
@@ -187,7 +187,7 @@ pub fn decode_instruction(bytes: &[u8]) -> Result<DecodedInstruction, DecodeErro
             let mod_bits = modrm >> 6;
             let rm = (modrm & 0x07) | (if rex.b { 0x08 } else { 0 });
 
-            let modrm_len = modrm_displacement_length(mod_bits, rm as u8 & 0x07, &bytes[pos..])?;
+            let modrm_len = modrm_displacement_length(mod_bits, rm & 0x07, &bytes[pos..])?;
             pos += modrm_len;
 
             if mod_bits == 0b11 {
@@ -214,7 +214,7 @@ pub fn decode_instruction(bytes: &[u8]) -> Result<DecodedInstruction, DecodeErro
             let mod_bits = modrm >> 6;
             let rm = (modrm & 0x07) | (if rex.b { 0x08 } else { 0 });
 
-            let modrm_len = modrm_displacement_length(mod_bits, rm as u8 & 0x07, &bytes[pos..])?;
+            let modrm_len = modrm_displacement_length(mod_bits, rm & 0x07, &bytes[pos..])?;
             pos += modrm_len;
 
             if mod_bits == 0b11 {
@@ -241,7 +241,7 @@ pub fn decode_instruction(bytes: &[u8]) -> Result<DecodedInstruction, DecodeErro
             let mod_bits = modrm >> 6;
             let rm = (modrm & 0x07) | (if rex.b { 0x08 } else { 0 });
 
-            let modrm_len = modrm_displacement_length(mod_bits, rm as u8 & 0x07, &bytes[pos..])?;
+            let modrm_len = modrm_displacement_length(mod_bits, rm & 0x07, &bytes[pos..])?;
             pos += modrm_len;
 
             if mod_bits == 0b11 {
@@ -277,8 +277,7 @@ pub fn decode_instruction(bytes: &[u8]) -> Result<DecodedInstruction, DecodeErro
                     let mod_bits = modrm >> 6;
                     let rm = (modrm & 0x07) | (if rex.b { 0x08 } else { 0 });
 
-                    let modrm_len =
-                        modrm_displacement_length(mod_bits, rm as u8 & 0x07, &bytes[pos..])?;
+                    let modrm_len = modrm_displacement_length(mod_bits, rm & 0x07, &bytes[pos..])?;
                     pos += modrm_len;
 
                     if mod_bits == 0b11 {
